Move conversion functions to their own modules
