+++ docs/wfdb.rst
@@ -9,7 +9,7 @@ WFDB Records
 ---------------
 
 .. automodule:: wfdb
-    :members: rdrecord, rdheader, rdsamp, wrsamp, edf2mit, mit2edf, wav2mit, mit2wav, csv2mit
+    :members: rdrecord, rdheader, rdsamp, wrsamp
 
 .. autoclass:: wfdb.Record
     :members: wrsamp, adc, dac
