+++ wfdb/processing/__init__.py
@@ -14,3 +14,4 @@ from wfdb.processing.evaluate import (
 from wfdb.processing.hr import compute_hr, calc_rr, calc_mean_hr
 from wfdb.processing.peaks import find_peaks, find_local_peaks, correct_peaks
 from wfdb.processing.qrs import XQRS, xqrs_detect, gqrs_detect
+from wfdb.processing.filter import sigavg
